requirements.txt
python-telegram-bot>=20.0
aiosqlite
beautifulsoup4
httpx
apscheduler
python-dotenv
